+++ tests/test_generate_tag.py
@@ -25,3 +25,15 @@ def test_reference(url, title, tag):
 )
 def test_inline(url, title, tag):
     assert generate_tag(url, title, "inline") == tag
+
+
+@pytest.mark.parametrize(
+    "url,title,tag",
+    [
+        ("foo", "bar", "[TODO](foo)"),
+        ("foo", "", "[TODO](foo)"),
+        ("foo", None, "[TODO](foo)"),
+    ],
+)
+def test_slack(url, title, tag):
+    assert generate_tag(url, title, "slack") == tag